--- libs/landing/src/lib/footer/footer.component.html
@@ -1,12 +1,0 @@
-<footer fxLayout="column" fxLayoutAlign="start center" fxLayout.lt-md="column" fxLayoutAlign.lt-md="center center">
-  <a routerLink="/">
-    <app-logo></app-logo>
-  </a>
-  <h3>Proudly associated with the following partners</h3>
-  <span fxLayout="row wrap" fxLayoutAlign="space-evenly start" fxLayoutGap="80px">
-    <img asset="CNC_logo.webp" type="logo" alt="cnc">
-    <img src="/assets/logo/BPIfrance.png" alt="BPI France logo"/>
-    <img src="/assets/logo/creativemediaeurope.png" alt="creatiev media europe logo"/>
-    <img asset="IFCIC_logo.webp" type="logo" alt="ifcic">
-  </span>
-</footer>
