--- libs/landing/src/lib/footer/footer.component.scss
@@ -1,21 +1,0 @@
-footer {
-  margin: 48px 0;
-  text-align: center;
-  width: 100%;
-
-  a {
-    margin-bottom: 56px;
-  }
-
-  img {
-    height: 40px;
-  }
-
-  app-logo img {
-    height: 48px;
-  }
-
-  h3 {
-    margin-bottom: 48px;
-  }
-}
