--- libs/landing/src/lib/header/header.component.html
@@ -1,10 +1,0 @@
-<header
-  fxLayout="column"
-  fxLayoutAlign="center start"
-  fxLayoutAlign.xs="center center">
-  <ng-content></ng-content>
-  <a mat-flat-button color="accent" routerLink="/auth/connexion">
-    <span>Join now</span>
-    <mat-icon svgIcon="arrow_forward"></mat-icon>
-  </a>
-</header>
