--- libs/landing/src/lib/header/header.component.scss
@@ -1,21 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-:host {
-  display: block;
-  padding-top: calc(var(--toolbarHeight) + 100px);
-  padding-bottom: 100px;
-  padding-left: 20%;
-  padding-right: 20%;
-  background-image: url('/assets/images/banner_header.png');
-  background-size: cover;
-}
-
-a {
-  margin-top: 80px;
-}
-
-@include layout-bp(xs) {
-  .offer {
-    font-size: 30px;
-  }
-}
