--- libs/landing/src/lib/header/header.component.ts
@@ -1,10 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-
-@Component({
-  selector: 'landing-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class LandingHeaderComponent {
-}
