--- libs/landing/src/lib/how-it-works/how-it-works.component.html
@@ -1,52 +1,0 @@
-<h1 class="mat-display-1">How it works?</h1>
-<mat-tab-group animationDuration="1000ms" selectedIndex="0">
-  <mat-tab [label]="tabNames[0]">
-    <article class="perspective-box" fxLayout="row" fxLayoutGap="24px" fxLayoutAlign="space-between start" fxLayout.lt-lg="column">
-      <img class="perspective-image" fxHide.xs [asset]="tabTitle[0].imgAsset" alt="seller image">
-      <div>
-        <h3>{{ tabTitle[0].title }}</h3>
-        <ng-container *ngIf="tabTitle[0].description as description">
-          <p class="mat-body-1">{{ description }}</p>
-        </ng-container>
-      </div>
-    </article>
-    <h5>Features</h5>
-    <article fxLayout="row wrap" fxLayoutAlign="space-between" fxLayout.lt-md="column">
-      <mat-card *ngFor="let feature of sellerFeatures">
-        <img class="feature-image" [asset]="feature.imgAsset" alt="seller feature image">
-        <div>
-          <h3>{{ feature.title }}</h3>
-          <p class="mat-body-1">{{ feature.description }}</p>
-        </div>
-      </mat-card>
-    </article>
-  </mat-tab>
-  <mat-tab [label]="tabNames[1]">
-    <article class="perspective-box" fxLayout="row" fxLayoutGap="24px" fxLayoutAlign="space-between start" fxLayout.lt-lg="column">
-      <img class="perspective-image" fxHide.xs [asset]="tabTitle[1].imgAsset" alt="buyer image">
-      <div>
-        <h3>{{ tabTitle[1].title }}</h3>
-        <ng-container *ngIf="tabTitle[1].description as description">
-          <p class="mat-body-1">{{ description }}</p>
-        </ng-container>
-      </div>
-    </article>
-    <h5>Features</h5>
-    <article fxLayout="row wrap" fxLayoutAlign="space-between" fxLayout.lt-md="column">
-      <mat-card *ngFor="let feature of buyerFeatures">
-        <img class="feature-image" [asset]="feature.imgAsset" alt="buyer feature image">
-        <div>
-          <h3>{{ feature.title }}</h3>
-          <p class="mat-caption">{{ feature.description }}</p>
-        </div>
-      </mat-card>
-    </article>
-  </mat-tab>
-</mat-tab-group>
-<section fxLayout="column" fxLayoutAlign="start center" class="join-section" fxLayoutGap="40px">
-  <p class="mat-display-2 interested">Interested? Join now!</p>
-  <a mat-flat-button color="accent" routerLink="/auth/connexion">
-    <span>Sign up</span>
-    <mat-icon svgIcon="arrow_forward"></mat-icon>
-  </a>
-</section>
