--- libs/landing/src/lib/how-it-works/how-it-works.component.scss
@@ -1,53 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-:host {
-  display: block;
-  margin: 48px 20% 0;
-
-  mat-card {
-    width: 49%;
-    margin-bottom: 24px;
-    text-align: center;
-  }
-}
-
-h1.mat-display-1 {
-  margin-bottom: 16px;
-}
-
-h3 {
-  margin-bottom: 8px;
-}
-
-h5 {
-  text-transform: uppercase;
-  margin-bottom: 8px;
-}
-
-.join-section {
-  margin: 80px 0;
-}
-
-.perspective-image {
-  width: 550px;
-}
-
-.perspective-box {
-  margin: 15px 0 24px 0;
-}
-
-.feature-image {
-  width: 150px;
-}
-
-@include layout-bp(xs) {
-  :host {
-    margin: 30px 20px;
-    mat-card {
-      width: 300px;
-    }
-  }
-  .interested {
-    font-size: 45px;
-  }
-}
