--- libs/landing/src/lib/how-it-works/how-it-works.component.ts
@@ -1,24 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-
-interface Display {
-  title: string,
-  description: string,
-  imgSrc?: string,
-  imgAsset?: string
-}
-
-@Component({
-  selector: '[tabNames] [tabTitle] [sellerFeatures] [buyerFeatures] landing-how-it-works',
-  templateUrl: './how-it-works.component.html',
-  styleUrls: ['./how-it-works.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class LandingHowItWorksComponent {
-  @Input() tabNames: string[];
-
-  @Input() tabTitle: Display[];
-
-  @Input() sellerFeatures: Display[];
-
-  @Input() buyerFeatures: Display[];
-}
