--- libs/landing/src/lib/learn-more/learn-more.component.html
@@ -1,61 +1,0 @@
-<mat-card id="background-card">
-  <article fxLayout="column" fxLayoutAlign="start center">
-    <h1>Want to learn more?</h1>
-    <h3>Schedule a demo with our team!</h3>
-    <p>
-      Program a personalized demonstration of the platform with a member of the Archipel team.
-    </p>
-    <p>
-      Please fill out the form and we will get back to you to choose the best time.
-    </p>
-  </article>
-  <form fxLayout="column" [formGroup]="form" (ngSubmit)="sendRequest(form)">
-    <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column"  fxLayoutGap="24px">
-      <mat-form-field appearance="outline">
-        <mat-label>First name</mat-label>
-        <input matInput type="text" placeholder="First name" formControlName="firstName" required />
-        <mat-error *ngIf="form.get('firstName').hasError('required')">
-          First name is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-      <mat-form-field appearance="outline">
-        <mat-label>Last name</mat-label>
-        <input matInput type="text" placeholder="Last name" formControlName="lastName" required />
-        <mat-error *ngIf="form.get('lastName').hasError('required')">
-          Last name is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-    </article>
-    <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column" fxLayoutGap="24px">
-      <mat-form-field appearance="outline">
-        <mat-label>Company name</mat-label>
-        <input matInput type="text" placeholder="Company name" formControlName="companyName" required />
-        <mat-error *ngIf="form.get('companyName').hasError('required')">
-          Company name is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-      <mat-form-field appearance="outline">
-        <mat-label>Activity</mat-label>
-        <mat-select placeholder="Activity" formControlName="role" required>
-          <mat-option *ngFor="let role of roles" [value]="role">{{ role | roleLabel }}</mat-option>
-        </mat-select>
-        <mat-error *ngIf="form.get('role').hasError('required')">
-          Activity is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-    </article>
-    <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column" fxLayoutGap="24px">
-      <mat-form-field appearance="outline">
-        <mat-label>Email</mat-label>
-        <input matInput type="email" placeholder="Email" formControlName="email" required />
-        <mat-error *ngIf="form.get('email').hasError('required')">
-            Email is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-      <mat-form-field appearance="outline">
-        <tel-input [form]="phoneNumber"  placeholder="Phone number"></tel-input>
-      </mat-form-field>
-    </article>
-    <button [disabled]="submitted" fxFlexAlign="center" mat-flat-button color="accent" type="submit">Send a demo request</button>
-  </form>
-</mat-card>
