--- libs/landing/src/lib/learn-more/learn-more.component.scss
@@ -1,32 +1,0 @@
-:host {
-  display: block;
-}
-
-mat-card#background-card {
-  border: none;
-  padding: 48px 0;
-}
-
-article {
-  width: 100%;
-}
-
-form {
-  padding-top: 24px;
-  margin: 0 20%;
-  .mat-form-field {
-    width: 100%;
-  }
-}
-
-button {
-  margin-top: 56px;
-}
-
-h3 {
-  margin-bottom: 8px;
-}
-
-p {
-  margin: 0;
-}
