--- libs/landing/src/lib/toolbar/toolbar.component.html
@@ -1,12 +1,0 @@
-<mat-toolbar fxLayout="row" fxLayoutAlign="space-between center">
-  <nav fxLayout="row" fxLayoutAlign="space-between center">
-    <a routerLink="/" class="logo">
-      <app-logo></app-logo>
-    </a>
-    <a fxHide.lt-md mat-stroked-button [color]="color" href="#demo" id="demo-link">Schedule a demo</a>
-  </nav>
-  <div fxLayout="row" fxLayoutAlign="space-around center" fxLayoutGap="16px">
-    <a test-id="login" mat-button [color]="color" routerLink="/auth/connexion" fragment="login">Log in</a>
-    <a test-id="signup" mat-button [color]="color" routerLink="/auth/connexion" fragment="signup">Sign up</a>
-  </div>
-</mat-toolbar>
