--- libs/landing/src/lib/toolbar/toolbar.component.scss
@@ -1,13 +1,0 @@
-.logo {
-  margin-right: 96px;
-}
-
-#demo-link {
-  letter-spacing: 0.5px;
-  line-height: 24px;
-  font-weight: 600;
-}
-
-mat-toolbar {
-  padding: 0 24px;
-}
